pyproject.toml
[project]
name = "vercel-fastapi-starter"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.104.0",
  "python-dotenv>=1.0.0",
  "pydantic>=2.0.0",
  "pydantic-ai>=0.0.14",
  "httpx>=0.25.0",
]
